+++ src/components/side-panel/coline-side-panel/coline-section.tsx
@@ -1,8 +1,8 @@
 import React from 'react';
+import { useDispatch } from 'react-redux';
 import { useAppSelector } from '../../../redux';
 import InterchangeCard from '../interchange/interchange-card';
-import { useDispatch } from 'react-redux';
-import { addColineColor, removeColineColor, updateColineColor } from '../../../redux/param/action';
+import { addColineColor, removeColineColor, updateColineColor } from '../../../redux/param/coline-action';
 import { InterchangeInfo } from '../../../constants/constants';
 
 export default function ColineSection() {
@@ -10,6 +10,9 @@ export default function ColineSection() {
 
     const selectedColine = useAppSelector(state => state.app.selectedColine);
     const { coline } = useAppSelector(state => state.param);
+
+    // return empty when selectedColine is invalid
+    if (selectedColine === undefined || selectedColine >= coline.length) return <></>;
 
     const handleAdd = () => (info: InterchangeInfo) => {
         dispatch(addColineColor(selectedColine, info));
@@ -23,7 +26,7 @@ export default function ColineSection() {
         dispatch(updateColineColor(selectedColine, i, info));
     };
 
-    // TODO: support mutiple coline color in one segement
+    // TODO: support multiple coline colors in one segment
     return (
         <InterchangeCard
             interchangeList={coline.at(selectedColine)!.colors}
