+++ src/redux/param/action.ts
@@ -629,50 +629,3 @@ export const removeStationService = (stationId: string, service: Services) => {
         }
     };
 };
-
-/**
- *
- * @param colineIndex
- * @param interchangeInfo
- */
-export const addColineColor = (colineIndex: number, interchangeInfo: InterchangeInfo) => {
-    return (dispatch: AppDispatch, getState: () => RootState) => {
-        const colineInfo = getState().param.coline;
-
-        colineInfo[colineIndex].colors = [...colineInfo[colineIndex].colors, interchangeInfo];
-
-        dispatch(setColineBulk(colineInfo));
-    };
-};
-
-export const removeColineColor = (colineIndex: number, interchangeIndex: number) => {
-    return (dispatch: AppDispatch, getState: () => RootState) => {
-        const colineInfo = getState().param.coline;
-
-        if (colineInfo.length > colineIndex && colineInfo[colineIndex].colors.length > interchangeIndex) {
-            colineInfo[colineIndex].colors = colineInfo[colineIndex].colors.filter(
-                (_, intIdx) => intIdx !== interchangeIndex
-            );
-
-            dispatch(setColineBulk(colineInfo));
-        }
-    };
-};
-
-export const updateColineColor = (colineIndex: number, interchangeIndex: number, interchangeInfo: InterchangeInfo) => {
-    return (dispatch: AppDispatch, getState: () => RootState) => {
-        const colineInfo = getState().param.coline;
-
-        if (colineInfo.length > colineIndex && colineInfo[colineIndex].colors.length > interchangeIndex) {
-            colineInfo[colineIndex].colors = colineInfo[colineIndex].colors.map((int, intIdx) =>
-                intIdx === interchangeIndex
-                    ? ([0, 1, 2, 3, 4, 5].map(i =>
-                          interchangeInfo[i] === undefined ? int[i] : interchangeInfo[i]
-                      ) as InterchangeInfo)
-                    : int
-            );
-
-            dispatch(setColineBulk(colineInfo));
-        }
-    };
-};
